Improve speed with faster window lookup
Window lookup used to be O(n) by looping through all the suppported
programs in bindings then looking if any of their window IDs matched
the current window ID. This was previously done since there wasn't a
way to get the class name of the current window and map it to the
custom keybindings.

I've changed it use xprop to get the instance name of the current window
which allows for O(1) lookup for bindings. This also required changing
some key names in bindings to match the new instance names from xprop.
Nice side effect of no longer dealing with bytestring.

The performance of O(n) typically shouldn't matter with such a small
dictionary of key bindings. But there was a noticable lag from key press
to command running because we'd have to wait for python to do run a
subprocess for xdotool for every program which made it slow.